✨ feat(sidebar): 将断开全部活跃连接入口移至侧边栏工具栏
