+++ apps/desktop/src/components/sidebar/ConnectionTree.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { ref, shallowRef, computed, nextTick, watch, provide, onMounted, onUnmounted, type Component, type ComponentPublicInstance, type CSSProperties } from "vue";
 import { useI18n } from "vue-i18n";
-import { Search, X, ListFilter, ListOrdered, ArrowDownAZ, ArrowUpZA, CircleDot, Crosshair, Server, Database, FolderTree, Table2, Eye, RotateCcw, Loader2, Unplug } from "@lucide/vue";
+import { Search, X, ListFilter, ListOrdered, ArrowDownAZ, ArrowUpZA, CircleDot, Crosshair, Server, Database, FolderTree, Table2, Eye, RotateCcw } from "@lucide/vue";
 import { useConnectionStore } from "@/stores/connectionStore";
 import { useQueryStore } from "@/stores/queryStore";
 import { useSavedSqlStore } from "@/stores/savedSqlStore";
@@ -82,7 +82,6 @@ import { alignedSidebarCommentLabelWidths, isSidebarCommentAlignableNode, sideba
 import { formatSidebarObjectStorage, sidebarTableStorageScopes, supportsSidebarTableStorage } from "@/lib/sidebar/sidebarDatabaseStorage";
 import { sidebarScrollbarGeometry as calculateSidebarScrollbarGeometry } from "@/lib/sidebar/sidebarScrollbar";
 import { createSidebarLayoutMonitor, type SidebarExpandedConnectionInfo } from "@/lib/sidebar/sidebarLayoutMonitor";
-import { disconnectSidebarConnections } from "@/lib/sidebar/sidebarConnectionDisconnect";
 import { compileSearchRegex } from "@/lib/common/searchPattern";
 
 const { t } = useI18n();
@@ -97,7 +96,6 @@ const regexMode = ref(false);
 const sidebarSearchLoadingTracker = createSidebarSearchLoadingTracker();
 const isSidebarSearchLoading = ref(false);
 const showConnectedConnectionsOnly = ref(false);
-const isDisconnectingAllActiveConnections = ref(false);
 const searchInputRef = ref<HTMLInputElement>();
 const rootRef = ref<HTMLElement>();
 const sidebarRootStyle = computed<Record<string, string>>(() => ({ fontSize: `${settingsStore.editorSettings.sidebarFontSize}px` }));
@@ -224,31 +222,6 @@ watch(
   },
   { flush: "sync" },
 );
-
-async function disconnectAllActiveConnections() {
-  if (isDisconnectingAllActiveConnections.value) return;
-  const connectionIds = [...store.connectedIds];
-  if (!connectionIds.length) {
-    showConnectedConnectionsOnly.value = false;
-    return;
-  }
-
-  isDisconnectingAllActiveConnections.value = true;
-  try {
-    const result = await disconnectSidebarConnections(connectionIds, (connectionId) => store.disconnect(connectionId));
-    if (!result.failed) {
-      toast(t("connection.disconnectedSelected", { count: connectionIds.length }), 2000);
-    } else if (result.succeeded > 0) {
-      toast(t("connection.disconnectSelectedPartial", { succeeded: result.succeeded, failed: result.failed }), 5000);
-    } else {
-      const message = result.firstError instanceof Error ? result.firstError.message : String(result.firstError);
-      toast(t("connection.saveFailed", { message }), 5000);
-    }
-  } finally {
-    isDisconnectingAllActiveConnections.value = false;
-    if (store.connectedIds.size === 0) showConnectedConnectionsOnly.value = false;
-  }
-}
 
 function refreshActiveSidebarTableSearches() {
   if (isTreeSearchFiltering.value) return;
@@ -2575,13 +2548,6 @@ defineExpose({ focusSearch, createNewGroup, collapseAllTreeNodes, locateTabInSid
         </div>
       </div>
     </CustomContextMenu>
-    <div v-if="showConnectedConnectionsOnly && store.connectedIds.size > 0" class="shrink-0 border-t border-border bg-background px-2 py-2">
-      <Button type="button" variant="outline" size="sm" class="h-7 w-full justify-center gap-1.5 text-xs" :disabled="isDisconnectingAllActiveConnections" @click="disconnectAllActiveConnections">
-        <Loader2 v-if="isDisconnectingAllActiveConnections" class="h-3.5 w-3.5 animate-spin" />
-        <Unplug v-else class="h-3.5 w-3.5" />
-        {{ t("sidebar.disconnectAllActiveConnections") }}
-      </Button>
-    </div>
     <SidebarDdlViewDialog
       v-if="sidebarDdlTarget"
       v-model:open="sidebarDdlOpen"
